# Switch store code from `Option<Digest>` to `Fingerprint`