--- a/src/rust/engine/fs/store/src/lib.rs
+++ b/src/rust/engine/fs/store/src/lib.rs
@@ -58,7 +58,7 @@ use fs::{
 };
 use futures::future::{self, BoxFuture, Either, FutureExt, TryFutureExt};
 use grpc_util::prost::MessageExt;
-use hashing::Digest;
+use hashing::{Digest, Fingerprint};
 use parking_lot::Mutex;
 use prost::Message;
 use protos::gen::build::bazel::remote::execution::v2 as remexec;
@@ -265,9 +265,11 @@ impl RemoteStore {
           if let Some(f_remote) = f_remote {
             f_remote(bytes.clone())?;
           }
+          let digest = Digest::of_bytes(&bytes);
           local_store
-            .store_bytes(entry_type, None, bytes, true)
-            .await?
+            .store_bytes(entry_type, digest.hash, bytes, true)
+            .await?;
+          digest
         } else {
           assert!(f_remote.is_none());
           // TODO(#18048): choose a file that can be plopped into the local store directly, when
@@ -444,10 +446,12 @@ impl Store {
     bytes: Bytes,
     initial_lease: bool,
   ) -> Result<Digest, String> {
+    let digest = Digest::of_bytes(&bytes);
     self
       .local
-      .store_bytes(EntryType::File, None, bytes, initial_lease)
-      .await
+      .store_bytes(EntryType::File, digest.hash, bytes, initial_lease)
+      .await?;
+    Ok(digest)
   }
 
   ///
@@ -458,9 +462,9 @@ impl Store {
   ///
   pub async fn store_file_bytes_batch(
     &self,
-    items: Vec<(Option<Digest>, Bytes)>,
+    items: Vec<(Fingerprint, Bytes)>,
     initial_lease: bool,
-  ) -> Result<Vec<Digest>, String> {
+  ) -> Result<(), String> {
     self
       .local
       .store_bytes_batch(EntryType::File, items, initial_lease)
@@ -560,7 +564,7 @@ impl Store {
         if cfg!(debug_assertions) {
           protos::verify_directory_canonical(d.digest(), &directory).unwrap();
         }
-        directories.push((Some(d.digest()), directory.to_bytes()))
+        directories.push((d.digest().hash, directory.to_bytes()))
       }
       directory::Entry::File(_) => (),
       directory::Entry::Symlink(_) => (),
@@ -568,11 +572,13 @@ impl Store {
 
     // Then store them as a batch.
     let local = self.local.clone();
-    let digests = local
+    let root = &directories[0];
+    let top_digest = Digest::new(root.0, root.1.len());
+    local
       .store_bytes_batch(EntryType::Directory, directories, initial_lease)
       .await?;
 
-    Ok(DirectoryDigest::new(digests[0], tree))
+    Ok(DirectoryDigest::new(top_digest, tree))
   }
 
   ///
@@ -586,10 +592,12 @@ impl Store {
     initial_lease: bool,
   ) -> Result<Digest, String> {
     let local = self.local.clone();
-    let digest = local
+    let bytes = directory.to_bytes();
+    let digest = Digest::of_bytes(&bytes);
+    local
       .store_bytes(
         EntryType::Directory,
-        None,
+        digest.hash,
         directory.to_bytes(),
         initial_lease,
       )