--- a/src/rust/engine/fs/store/src/local.rs
+++ b/src/rust/engine/fs/store/src/local.rs
@@ -251,27 +251,23 @@ impl ByteStore {
   }
 
   ///
-  /// Store the given data in a single pass, optionally using the given Digest. Prefer `Self::store`
+  /// Store the given data in a single pass, using the given Fingerprint. Prefer `Self::store`
   /// for values which should not be pulled into memory, and `Self::store_bytes_batch` when storing
   /// multiple values at a time.
   ///
   pub async fn store_bytes(
     &self,
     entry_type: EntryType,
-    digest: Option<Digest>,
+    fingerprint: Fingerprint,
     bytes: Bytes,
     initial_lease: bool,
-  ) -> Result<Digest, String> {
+  ) -> Result<(), String> {
     let dbs = match entry_type {
       EntryType::Directory => self.inner.directory_dbs.clone(),
       EntryType::File => self.inner.file_dbs.clone(),
     };
-    let len = bytes.len();
-    let fingerprint = dbs?
-      .store_bytes(digest.map(|d| d.hash), bytes, initial_lease)
-      .await?;
-
-    Ok(Digest::new(fingerprint, len))
+    dbs?.store_bytes(fingerprint, bytes, initial_lease).await?;
+    Ok(())
   }
 
   ///
@@ -283,37 +279,17 @@ impl ByteStore {
   pub async fn store_bytes_batch(
     &self,
     entry_type: EntryType,
-    items: Vec<(Option<Digest>, Bytes)>,
+    items: Vec<(Fingerprint, Bytes)>,
     initial_lease: bool,
-  ) -> Result<Vec<Digest>, String> {
+  ) -> Result<(), String> {
     let dbs = match entry_type {
       EntryType::Directory => self.inner.directory_dbs.clone(),
       EntryType::File => self.inner.file_dbs.clone(),
     };
-    // NB: False positive: we do actually need to create the Vec here, since `items` will move
-    // before we use `lens`.
-    #[allow(clippy::needless_collect)]
-    let lens = items
-      .iter()
-      .map(|(_, bytes)| bytes.len())
-      .collect::<Vec<_>>();
-    let fingerprints = dbs?
-      .store_bytes_batch(
-        items
-          .into_iter()
-          .map(|(d, bytes)| (d.map(|d| d.hash), bytes))
-          .collect(),
-        initial_lease,
-      )
-      .await?;
 
-    Ok(
-      fingerprints
-        .into_iter()
-        .zip(lens.into_iter())
-        .map(|(f, len)| Digest::new(f, len))
-        .collect(),
-    )
+    dbs?.store_bytes_batch(items, initial_lease).await?;
+
+    Ok(())
   }
 
   ///