Prefactor for #18153, and overall cleanup.

The only reason I could think of wanting `Option<>` is if we saved ourselves an extra pass of the data when consuming the bytes, but the ultimate consumer was `src/rust/engine/sharded_lmdb/src/lib.rs` which just hashed the bytes (in a sync thread too!).