@thejcannon 2023-02-24
I don't see a whole lot of variance...

Before:
```
josh@cephandrius:~/work/pants$ hyperfine 'pants --no-pantsd --rule-threads-core=4 --rule-threads-max=16 lint ::'
Benchmark 1: pants --no-pantsd --rule-threads-core=4 --rule-threads-max=16 lint ::
  Time (mean ± σ):     14.528 s ±  0.405 s    [User: 14.661 s, System: 2.143 s]
  Range (min … max):   14.028 s … 15.563 s    10 runs
```

After:
```
josh@cephandrius:~/work/pants$ hyperfine 'pants --no-pantsd --rule-threads-core=4 --rule-threads-max=16  lint ::'
Benchmark 1: pants --no-pantsd --rule-threads-core=4 --rule-threads-max=16  lint ::
  Time (mean ± σ):     14.515 s ±  0.320 s    [User: 14.682 s, System: 2.170 s]
  Range (min … max):   13.927 s … 15.127 s    10 runs


josh@cephandrius:~/work/pants$ hyperfine 'pants --no-pantsd --rule-threads-core=8 --rule-threads-max=16  lint ::'
Benchmark 1: pants --no-pantsd --rule-threads-core=8 --rule-threads-max=16  lint ::
  Time (mean ± σ):     14.505 s ±  0.271 s    [User: 15.587 s, System: 2.232 s]
  Range (min … max):   14.319 s … 15.223 s    10 runs

josh@cephandrius:~/work/pants$ hyperfine 'pants --no-pantsd --rule-threads-core=12 --rule-threads-max=16  lint ::'
Benchmark 1: pants --no-pantsd --rule-threads-core=12 --rule-threads-max=16  lint ::
  Time (mean ± σ):     14.556 s ±  0.341 s    [User: 15.814 s, System: 2.249 s]
  Range (min … max):   14.315 s … 15.475 s    10 runs
```

so maybe switching to half-and-half would better?